Added validation routines to user managment in Admin portal

+++ AGNES/My Project/Settings.Designer.vb
@@ -100,9 +100,9 @@ Partial Friend NotInheritable Class MySettings
     <Global.System.Configuration.UserScopedSettingAttribute(),  _
      Global.System.Diagnostics.DebuggerNonUserCodeAttribute(),  _
      Global.System.Configuration.DefaultSettingValueAttribute("0")>  _
-    Public Property UserID() As Integer
+    Public Property UserID() As Long
         Get
-            Return CType(Me("UserID"),Integer)
+            Return CType(Me("UserID"),Long)
         End Get
         Set
             Me("UserID") = value
